Feature/refatorizacion ajustes varios

+++ ms-ejercicios/src/main/java/com/example/ms_ejercicios/service/EjercicioServiceImpl.java
@@ -1,7 +1,6 @@
 package com.example.ms_ejercicios.service;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
@@ -38,6 +37,10 @@ public class EjercicioServiceImpl implements EjercicioService {
 
     @Override
     public void eliminarPorId(Long id) {
+       if (!ejercicioRepository.existsById(id)) {
+           throw new RuntimeException("No se encontró un ejercicio con el id: " + id);
+       }
+
        ejercicioRepository.deleteById(id);
     }
 
@@ -47,8 +50,9 @@ public class EjercicioServiceImpl implements EjercicioService {
     }
 
     @Override
-    public Optional<Ejercicio> buscarPorNombre(String nombre) {
-         return ejercicioRepository.findByNombreEjercicio(nombre);
+    public Ejercicio buscarPorNombre(String nombre) {
+         return ejercicioRepository.findByNombreEjercicio(nombre)
+             .orElseThrow(() -> new RuntimeException("ejercicio no encontrado"));
     }
 
 }
